Optimize code structure and fix a minor bug
* Add DragHandle, a component for easing drag handling
* Fix card doesn't redraw after layout reset, causing curve to stretch
